oidc-risk-module-main/scikit_fuzzy/claimsdata/requirements.txt
Django>=5.0
psycopg2>=2.8
psycopg
djangorestframework
